implemented player movement in ECS (without collision)

+++ core/src/main/java/ua/leonidius/coldline/entity/components/MovementComponent.kt
@@ -5,5 +5,5 @@ import com.badlogic.gdx.math.Vector2
 
 data class MovementComponent(
     val velocity: Vector2 = Vector2(0F, 0F),
-    var speed: Int = 60 * 2
+    var speed: Float = 60 * 2F
 ): Component
